Separate flags, add VolumeFlag, refactor Volume and VolumeMount types
